Create style.js
