+++ client/src/assets/styles/style.js
@@ -1,0 +1,128 @@
+const rowStyle = {
+    display: 'flex',
+    justifyContent: 'space-between',
+    flexGrow: 1,
+    alignItems: 'center', 
+    color: 'black', 
+  };
+
+  const textContainerStyle = {
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'center',
+    alignItems: 'center',
+  };
+
+  const homeStyle = {
+    display: 'block',
+    flexDirection: 'column',
+    justifyContent: 'center',
+    alignItems: 'center',
+    textAlign: 'center',
+    height: '100%',
+    color: 'white',
+  };
+
+  const homeStyleParapgraph = {
+    color: 'white', 
+    fontWeight: 'bold',
+    fontSize: '25px',
+  };
+
+  const centeredContentStyle = {
+    display: 'flex',
+    flexDirection: 'column',
+    justifyContent: 'center',
+    alignItems: 'center',
+  };
+  
+  const h1Style = {
+    color: '#8B1D1D',
+    display: 'flex',
+    alignItems: 'center', 
+    justifyContent: 'center',
+  };
+
+  const h2Style = {
+    color: '#8B1D1D',
+    display: 'flex',
+    alignItems: 'center', 
+    justifyContent: 'center',
+    marginBottom: 100,
+  };
+  
+  const pStyle = {
+    color: '#C0CA33', 
+    fontSize: '25px',
+  };
+
+  const imgStyle = {
+    zIndex: '999', 
+    width: '150px',
+  };
+
+  const buttonTextStyle = {
+    color: 'black',
+    fontWeight: 'bold', 
+  };
+
+  const containerStyleInline = {
+    display: 'flex',
+    alignItems: 'center', 
+    justifyContent: 'center',
+    color: '#8B1D1D', 
+  };
+  
+  const iconStyle = {
+    fontSize: 100,
+    marginRight: 15,
+    marginLeft: 15,
+    marginBottom: -5,
+  };
+
+  const iconStyleLine = {
+    fontSize: 100,
+    marginRight: 2,
+    height: 100,
+    marginBottom: -4,
+  };
+
+  const paragraphContainerStyle = {
+    display: 'flex',
+    justifyContent: 'center',
+  };
+
+  const pStyle123 = {
+    color: '#C0CA33', 
+    fontSize: '25px',
+    marginLeft: '100px',
+    marginRight: '100px',
+    marginBottom: '100px',
+  };
+
+  const rond = {
+    zIndex: '999',
+    width: '150px',
+    marginLeft: '150px',
+    marginRight: '150px',
+    marginBottom: '100px',
+
+  };
+
+  const pRond = {
+    color: '#C0CA33', 
+    marginLeft: '-100px',
+    fontSize: '20px',
+    marginBottom: '100px',
+  };
+
+  const footerBot = {
+    marginBottom: '100px',
+    color: '#8B1D1D',
+    fontSize: '30px',
+    display: 'flex',
+    alignItems: 'center', 
+    justifyContent: 'center',
+  };
+
+export { footerBot, pRond, rond, pStyle123, paragraphContainerStyle ,iconStyleLine ,iconStyle , containerStyleInline, rowStyle, textContainerStyle, homeStyle, homeStyleParapgraph, centeredContentStyle, h1Style, h2Style, pStyle, imgStyle, buttonTextStyle };
